# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Nervia API Documentation",
    "version": "1.0.0",
    "description": "API for Nervia Web Clipper and Knowledge Graph integration"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [],
  "paths": {
...
    "/api/ai/process": {
      "post": {
        "description": "AI processing for the knowledge graph. Modes - suggest_connections (find related neurons + category) or analyze_link (summary, tags, connections for a new link).",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mode": {
                    "type": "string",
                    "enum": [
                      "suggest_connections",
                      "analyze_link"
                    ]
                  },
                  "newNode": {
                    "type": "object"
                  },
                  "existingNodes": {
                    "type": "array"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Processed result (description/summary, connections, group_id, tags)."
          },
          "401": {
            "description": "Unauthorized."
          },
          "429": {
            "description": "Rate limit (RATE_LIMIT, retryAfterSeconds)."
          },
          "500": {
            "description": "API key missing or server error."
          }
        }
      }
    },
...
  },
...
}